LayoutTests/media/wireless-playback-media-player/has-audio.html

<!DOCTYPE html><!-- webkit-test-runner [ WirelessPlaybackMediaPlayerEnabled=true ] -->
<html>
    <head>
        <script src='../media-file.js'></script>
        <script src='../video-test.js'></script>
        <script>
            var route;
            var video;

            async function start()
            {
                if (window.internals)
                    internals.mockMediaDeviceRouteController.enabled = true;

                route = internals.mockMediaDeviceRouteController.createMockMediaDeviceRoute();
                run(`internals.mockMediaDeviceRouteController.activateRoute(route)`);

                const urlPromise = new Promise((resolve) => {
                    route.setURLCallback(async (url) => {
                        resolve(url);
                    });
                });

                video = document.getElementsByTagName('video')[0];
                run(`video.src = findMediaFile('video', '../content/test')`);
                run(`video.play()`);
                await urlPromise;

                const canplaythroughPromise = waitFor(video, 'canplaythrough');
                run(`route.ready = true`);
                await canplaythroughPromise;

                run(`route.audioOptions = [{ displayName: 'Audio', identifier: 'audio' }]`);
                await testExpectedEventually('internals.mediaUsageState(video).hasAudio', true);

                run(`route.audioOptions = []`);
                await testExpectedEventually('internals.mediaUsageState(video).hasAudio', false);

                endTest();
            }
        </script>
    </head>
    <body onload='start()'>
        <video controls></video>
        <p>Test that toggling a MediaDeviceRoute's audioOptions propagates to the video element's hasAudio state.</p>
    </body>
</html>
